pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hyperheadset"
version = "0.1.0"
description = "Astro A50 base station HID stats client"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Hyper" }]
dependencies = [
  "hidapi>=0.14.0"
]

[project.scripts]
hyperheadset = "hyperheadset.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
